+++ docker/entrypoint.sh
@@ -4,7 +4,14 @@
 
 echo "Welcome to Machinaris!"
 
-# Start selected forks
+# v0.6.0 upgrade check guard - only allow single blockchain per container
+if [[ "${blockchains}" == "chia,flax" ]]; then
+  echo "Only one blockchain allowed per container in Machinaris."
+  echo "Please remove second value from 'blockchains' variable and restart."
+  return 1
+fi
+
+# Start only selected fork
 for fork in ${blockchains//,/ }; do
   /usr/bin/bash /machinaris/scripts/forks/${fork}_launch.sh
 done