+++ scripts/chiadog_install.sh
@@ -3,7 +3,7 @@
 # Installs Chiadog for log monitoring and alerting
 #
 
-CHIADOG_BRANCH=main
+CHIADOG_BRANCH=dev
 
 echo 'Installing Chiadog...'