+++ scripts/forks/nchain_launch.sh
@@ -57,24 +57,25 @@ elif [[ ${mode} =~ ^harvester.* ]]; then
     echo "A farmer peer address and port are required."
     exit
   else
-    if [ ! -f /root/.chia/ext9/farmer_ca/chia_ca.crt ]; then
-      mkdir -p /root/.chia/ext9/farmer_ca
+    if [ ! -f /root/.chia/farmer_ca/chia_ca.crt ]; then
+      mkdir -p /root/.chia/farmer_ca
       response=$(curl --write-out '%{http_code}' --silent http://${controller_host}:8929/certificates/?type=nchain --output /tmp/certs.zip)
       if [ $response == '200' ]; then
-        unzip /tmp/certs.zip -d /root/.chia/ext9/farmer_ca
+        unzip /tmp/certs.zip -d /root/.chia/farmer_ca
       else
         echo "Certificates response of ${response} from http://${controller_host}:8929/certificates/?type=nchain.  Try clicking 'New Worker' button on 'Workers' page first."
       fi
       rm -f /tmp/certs.zip 
     fi
-    if [ -f /root/.chia/ext9/farmer_ca/chia_ca.crt ]; then
-      chia init -c /root/.chia/ext9/farmer_ca 2>&1 > /root/.chia/log/init.log
+    if [ -f /root/.chia/farmer_ca/chia_ca.crt ]; then
+      chia init -c /root/.chia/farmer_ca 2>&1 > /root/.chia/ext9/log/init.log
       chmod 755 -R /root/.chia/ext9/config/ssl/ &> /dev/null
       chia init --fix-ssl-permissions > /dev/null 
     else
       echo "Did not find your farmer's certificates within /root/.chia/farmer_ca."
       echo "See: https://github.com/guydavis/machinaris/wiki/Workers#harvester"
     fi
+    echo "Configuring farmer peer at ${farmer_address}:${farmer_port}"
     chia configure --set-farmer-peer ${farmer_address}:${farmer_port}
     chia configure --enable-upnp false
     chia start harvester -r